pyproject.toml
[tool.poetry]
name = "certbot-dns-godaddy"
version = "0.1.1"
description = "A certbot plugin that implements letsencrypt dns wildcard support for godaddy using lexicon"
authors = ["Dustyn Gibson <miigotu@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/miigotu/certbot-dns-godaddy"

packages = [{include = "certbot_dns_godaddy.py"}]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = ">=3.6,<4.0"
acme = ">=0.31.0"
certbot = ">=0.31.0"
dns-lexicon = ">=3.2.3"
"zope.interface" = "^5.4.0"
urllib3 = "^1.26.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."certbot.plugins"]
"dns-godaddy" = "certbot_dns_godaddy:Authenticator"